base64: split up base64 files in plugin functions and encoding functions
base64.c contains only the plugin functions (get,set).
base64_functions.c contains the encoding and decoding functions, so that other plugins can include them easily.
