+++ src/plugins/base64/base64.h
@@ -14,13 +14,7 @@
 #include <kdbtypes.h>
 #include <stdio.h>
 
-#define ELEKTRA_PLUGIN_BASE64_PREFIX "@BASE64"
-#define ELEKTRA_PLUGIN_BASE64_ESCAPE "@"
-#define ELEKTRA_PLUGIN_BASE64_ESCAPE_CHAR '@'
-
-// encoding functions
-char * ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME_C, base64Encode) (const kdb_octet_t * input, const size_t inputLength);
-int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME_C, base64Decode) (const char * input, kdb_octet_t ** output, size_t * outputLength);
+#include "base64_functions.h"
 
 // kdb functions
 int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME_C, get) (Plugin * handle, KeySet * ks, Key * parentKey);
